functions/service_registry/requirements.txt
google-cloud-artifact-registry==1.*
